fastapi-workshop/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pydantic
jose
python-dotenv
pydantic-settings
